+++ src/main/java/es/in2/issuer/backend/oidc4vci/domain/service/TokenService.java
@@ -4,5 +4,5 @@ import es.in2.issuer.backend.oidc4vci.domain.model.TokenResponse;
 import reactor.core.publisher.Mono;
 
 public interface TokenService {
-    Mono<TokenResponse> generateTokenResponse(String grantType, String preAuthorizedCode, String txCode);
+    Mono<TokenResponse> generateTokenResponse(String grantType, String preAuthorizedCode, String txCode, String refreshToken);
 }