+++ src/main/java/es/in2/issuer/backend/oidc4vci/domain/service/impl/TokenServiceImpl.java
@@ -11,6 +11,7 @@ import es.in2.issuer.backend.shared.infrastructure.config.AppConfig;
 import es.in2.issuer.backend.shared.infrastructure.repository.CacheStore;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.jetbrains.annotations.NotNull;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Mono;
 
@@ -21,6 +22,7 @@ import java.util.NoSuchElementException;
 
 import static es.in2.issuer.backend.oidc4vci.domain.util.Constants.ACCESS_TOKEN_EXPIRATION_TIME_DAYS;
 import static es.in2.issuer.backend.shared.domain.util.Constants.GRANT_TYPE;
+import static es.in2.issuer.backend.shared.domain.util.Constants.REFRESH_TOKEN_GRANT_TYPE;
 
 @Slf4j
 @Service
@@ -36,20 +38,49 @@ public class TokenServiceImpl implements TokenService {
     private final AppConfig appConfig;
 
     @Override
-    public Mono<TokenResponse> generateTokenResponse(String grantType, String preAuthorizedCode, String txCode) {
+    public Mono<TokenResponse> generateTokenResponse(String grantType, String preAuthorizedCode, String txCode, String refreshToken) {
         log.debug("Generating token response for pre-authorized code: {}", preAuthorizedCode);
 
         return validateGrantType(grantType)
-                .then(validatePreAuthorizedCodeAndTxCode(preAuthorizedCode, txCode))
-                .then(generateTokens(preAuthorizedCode))
+                .then(validateByGrantType(grantType, preAuthorizedCode, txCode, refreshToken))
+                .then(Mono.defer(() -> generateTokens(preAuthorizedCode)))
                 .doOnSuccess(response -> log.debug("Token response generated successfully"))
                 .doOnError(error -> log.error("Error generating token response for code {}: {}", preAuthorizedCode, error.getMessage()));
     }
 
+    private @NotNull Mono<Void> validateByGrantType(String grantType, String preAuthorizedCode, String txCode, String refreshToken) {
+        return Mono.defer(() ->
+                REFRESH_TOKEN_GRANT_TYPE.equals(grantType)
+                        ? validateRefreshToken(refreshToken)
+                        : validatePreAuthorizedCodeAndTxCode(preAuthorizedCode, txCode));
+    }
+
     private Mono<Void> validateGrantType(String grantType) {
-        return GRANT_TYPE.equals(grantType)
+        return GRANT_TYPE.equals(grantType) || REFRESH_TOKEN_GRANT_TYPE.equals(grantType)
                 ? Mono.empty()
                 : Mono.error(new IllegalArgumentException("Invalid grant type: " + grantType));
+    }
+
+    private Mono<Void> validateRefreshToken(String refreshToken) {
+        return refreshTokenCacheStore
+                .get(refreshToken)
+                .doOnError(error -> log.error("Failed to retrieve refresh token"))
+                .onErrorMap(NoSuchElementException.class, ex -> new IllegalArgumentException("Invalid refresh token"))
+                .flatMap(credentialProcedureIdAndRefreshToken -> {
+                    if (credentialProcedureIdAndRefreshToken.refreshTokenJti().equals(refreshToken)) {
+                        long currentTime = Instant.now().getEpochSecond();
+                        if (currentTime < credentialProcedureIdAndRefreshToken.refreshTokenExpiresAt()) {
+                            return Mono.empty();
+                        } else {
+                            log.error("Refresh token expired ");
+                            return Mono.error(new IllegalArgumentException("Refresh token expired"));
+                        }
+                    } else {
+                        log.error("Invalid refresh token provided");
+                        return Mono.error(new IllegalArgumentException("Invalid refresh token"));
+                    }
+                })
+                .then(refreshTokenCacheStore.delete(refreshToken));
     }
 
     private Mono<Void> validatePreAuthorizedCodeAndTxCode(String preAuthorizedCode, String txCode) {
@@ -107,12 +138,13 @@ public class TokenServiceImpl implements TokenService {
 
         CredentialProcedureIdAndRefreshToken credentialProcedureIdAndRefreshToken =
                 CredentialProcedureIdAndRefreshToken.builder()
+                        .preAuthorizedCode(preAuthorizedCode)
                         .credentialProcedureId(credentialProcedureId)
                         .refreshTokenJti(refreshToken)
                         .refreshTokenExpiresAt(refreshTokenExpiresAt)
                         .build();
 
-        return refreshTokenCacheStore.add(preAuthorizedCode, credentialProcedureIdAndRefreshToken).then();
+        return refreshTokenCacheStore.add(refreshToken, credentialProcedureIdAndRefreshToken).then();
     }
 
     private TokenResponse buildTokenResponse(String accessToken, long accessTokenExpirationTime) {