+++ src/main/java/es/in2/issuer/backend/shared/domain/model/dto/CredentialProcedureIdAndRefreshToken.java
@@ -4,6 +4,7 @@ import lombok.Builder;
 
 @Builder
 public record CredentialProcedureIdAndRefreshToken(
+        String preAuthorizedCode,
         String credentialProcedureId,
         String refreshTokenJti,
         long refreshTokenExpiresAt) {