+++ src/test/java/es/in2/issuer/backend/oidc4vci/infrastructure/controller/TokenControllerTest.java
@@ -45,7 +45,8 @@ class TokenControllerTest {
         when(tokenService.generateTokenResponse(
                 grantType,
                 preAuthorizedCode,
-                txCode))
+                txCode
+                , ""))
                 .thenReturn(Mono.just(tokenResponse));
 
         webTestClient