+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v2.2.7](https://github.com/in2workspace/in2-issuer-api/releases/tag/v2.2.7)
+### Changed
+- Update refresh token.
+- Update deferred credential flow.
+
 ## [v2.2.6](https://github.com/in2workspace/in2-issuer-api/releases/tag/v2.2.6)
 ### Changed
 - Set vault's secret mounts as environment variable.